[CodingStandard] drop NoInterfaceOnAbstractClassFixer, rename test to specific class names
