+++ packages/CodingStandard/tests/Fixer/ControlStructure/RequireFollowedByAbsolutePathFixer/RequireFollowedByAbsolutePathFixerTest.php
@@ -6,7 +6,7 @@ use PhpCsFixer\Fixer\FixerInterface;
 use Symplify\CodingStandard\Fixer\ControlStructure\RequireFollowedByAbsolutePathFixer;
 use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 
-final class Test extends AbstractSimpleFixerTestCase
+final class RequireFollowedByAbsolutePathFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
      * @dataProvider provideFixCases()
