+++ packages/CodingStandard/tests/Fixer/Import/ImportNamespacedNameFixer/ImportNamespacedNameFixerTest.php
@@ -1,0 +1,46 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\CodingStandard\Tests\Fixer\Import\ImportNamespacedNameFixer;
+
+use PhpCsFixer\Fixer\FixerInterface;
+use Symplify\CodingStandard\Fixer\Import\ImportNamespacedNameFixer;
+use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
+
+final class ImportNamespacedNameFixerTest extends AbstractSimpleFixerTestCase
+{
+    /**
+     * @dataProvider provideFixCases()
+     */
+    public function testFix(string $expected, ?string $input = null): void
+    {
+        $this->doTest($expected, $input);
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function provideFixCases(): array
+    {
+        return [
+            # correct
+            [__DIR__ . '/correct/correct.php.inc'],
+            [__DIR__ . '/correct/correct2.php.inc'],
+            # wrong => fixed
+            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
+            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
+            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
+            [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc', ],
+            [__DIR__ . '/wrong/wrong5.php.inc', __DIR__ . '/fixed/fixed5.php.inc', ],
+            [__DIR__ . '/wrong/wrong6.php.inc', __DIR__ . '/fixed/fixed6.php.inc', ],
+            [__DIR__ . '/wrong/wrong7.php.inc', __DIR__ . '/fixed/fixed7.php.inc', ],
+            [__DIR__ . '/wrong/wrong8.php.inc', __DIR__ . '/fixed/fixed8.php.inc', ],
+            [__DIR__ . '/wrong/wrong9.php.inc', __DIR__ . '/fixed/fixed9.php.inc', ],
+            [__DIR__ . '/wrong/wrong10.php.inc', __DIR__ . '/fixed/fixed10.php.inc', ],
+        ];
+    }
+
+    protected function createFixer(): FixerInterface
+    {
+        return new ImportNamespacedNameFixer();
+    }
+}
