+++ packages/CodingStandard/tests/Fixer/Naming/PropertyNameMatchingTypeFixer/PropertyNameMatchingTypeFixerTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\CodingStandard\Tests\Fixer\Import\ImportNamespacedNameFixer;
+namespace Symplify\CodingStandard\Tests\Fixer\Naming\PropertyNameMatchingTypeFixer;
 
 use PhpCsFixer\Fixer\FixerInterface;
-use Symplify\CodingStandard\Fixer\Import\ImportNamespacedNameFixer;
+use Symplify\CodingStandard\Fixer\Naming\PropertyNameMatchingTypeFixer;
 use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 
-final class Test extends AbstractSimpleFixerTestCase
+final class PropertyNameMatchingTypeFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
      * @dataProvider provideFixCases()
@@ -23,24 +23,17 @@ final class Test extends AbstractSimpleFixerTestCase
     {
         return [
             # correct
-            [__DIR__ . '/correct/correct.php.inc'],
-            [__DIR__ . '/correct/correct2.php.inc'],
+            [__DIR__ . '/correct/correct.php.inc', ],
             # wrong => fixed
             [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
             [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
             [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
             [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc', ],
-            [__DIR__ . '/wrong/wrong5.php.inc', __DIR__ . '/fixed/fixed5.php.inc', ],
-            [__DIR__ . '/wrong/wrong6.php.inc', __DIR__ . '/fixed/fixed6.php.inc', ],
-            [__DIR__ . '/wrong/wrong7.php.inc', __DIR__ . '/fixed/fixed7.php.inc', ],
-            [__DIR__ . '/wrong/wrong8.php.inc', __DIR__ . '/fixed/fixed8.php.inc', ],
-            [__DIR__ . '/wrong/wrong9.php.inc', __DIR__ . '/fixed/fixed9.php.inc', ],
-            [__DIR__ . '/wrong/wrong10.php.inc', __DIR__ . '/fixed/fixed10.php.inc', ],
         ];
     }
 
     protected function createFixer(): FixerInterface
     {
-        return new ImportNamespacedNameFixer();
+        return new PropertyNameMatchingTypeFixer();
     }
 }
