+++ packages/CodingStandard/tests/Fixer/Php/ClassStringToClassConstantFixer/ClassStringToClassConstantFixerTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\CodingStandard\Tests\Fixer\Property\ArrayPropertyDefaultValueFixer;
+namespace Symplify\CodingStandard\Tests\Fixer\Php\ClassStringToClassConstantFixer;
 
 use PhpCsFixer\Fixer\FixerInterface;
-use Symplify\CodingStandard\Fixer\Property\ArrayPropertyDefaultValueFixer;
+use Symplify\CodingStandard\Fixer\Php\ClassStringToClassConstantFixer;
 use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 
-final class Test extends AbstractSimpleFixerTestCase
+final class ClassStringToClassConstantFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
      * @dataProvider provideFixCases()
@@ -26,13 +26,15 @@ final class Test extends AbstractSimpleFixerTestCase
             [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
             [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
             [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
-            [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc', ],
             # correct
-            [__DIR__ . '/correct/correct.php.inc', ], ];
+            ['<?php $form->addText(\'datetime\');'],
+            ['<?php $request->getParameter(\'exception\');'],
+            ['<?php $this->assertTrue(class_exists(\'\ApiGen\Reflection\Tests\ExtendingClass\'));'],
+        ];
     }
 
     protected function createFixer(): FixerInterface
     {
-        return new ArrayPropertyDefaultValueFixer();
+        return new ClassStringToClassConstantFixer();
     }
 }
