+++ packages/CodingStandard/tests/Fixer/Solid/FinalInterfaceFixer/FinalInterfaceFixerTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\CodingStandard\Tests\Fixer\Naming\PropertyNameMatchingTypeFixer;
+namespace Symplify\CodingStandard\Tests\Fixer\Solid\FinalInterfaceFixer;
 
 use PhpCsFixer\Fixer\FixerInterface;
-use Symplify\CodingStandard\Fixer\Naming\PropertyNameMatchingTypeFixer;
+use Symplify\CodingStandard\Fixer\Solid\FinalInterfaceFixer;
 use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 
-final class Test extends AbstractSimpleFixerTestCase
+final class FinalInterfaceFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
      * @dataProvider provideFixCases()
@@ -22,18 +22,20 @@ final class Test extends AbstractSimpleFixerTestCase
     public function provideFixCases(): array
     {
         return [
-            # correct
-            [__DIR__ . '/correct/correct.php.inc', ],
             # wrong => fixed
             [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
             [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
-            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
-            [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc', ],
+            # correct
+            [__DIR__ . '/correct/correct.php.inc'],
+            [__DIR__ . '/correct/correct2.php.inc'],
+            [__DIR__ . '/correct/correct3.php.inc'],
+            [__DIR__ . '/correct/correct4.php.inc'],
+            [__DIR__ . '/correct/correct5.php.inc'],
         ];
     }
 
     protected function createFixer(): FixerInterface
     {
-        return new PropertyNameMatchingTypeFixer();
+        return new FinalInterfaceFixer();
     }
 }
