--- packages/CodingStandard/tests/Fixer/Solid/FinalInterfaceFixer/Test.php
@@ -1,41 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\CodingStandard\Tests\Fixer\Solid\FinalInterfaceFixer;
-
-use PhpCsFixer\Fixer\FixerInterface;
-use Symplify\CodingStandard\Fixer\Solid\FinalInterfaceFixer;
-use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
-
-final class Test extends AbstractSimpleFixerTestCase
-{
-    /**
-     * @dataProvider provideFixCases()
-     */
-    public function testFix(string $expected, ?string $input = null): void
-    {
-        $this->doTest($expected, $input);
-    }
-
-    /**
-     * @return string[][]
-     */
-    public function provideFixCases(): array
-    {
-        return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
-            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
-            # correct
-            [__DIR__ . '/correct/correct.php.inc'],
-            [__DIR__ . '/correct/correct2.php.inc'],
-            [__DIR__ . '/correct/correct3.php.inc'],
-            [__DIR__ . '/correct/correct4.php.inc'],
-            [__DIR__ . '/correct/correct5.php.inc'],
-        ];
-    }
-
-    protected function createFixer(): FixerInterface
-    {
-        return new FinalInterfaceFixer();
-    }
-}
