--- packages/CodingStandard/tests/Fixer/Solid/NoInterfaceOnAbstractClassFixer/fixed/fixed.php.inc
@@ -1,9 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-
-abstract class SomeClass
-{
-
-}
