--- packages/CodingStandard/tests/Fixer/Solid/NoInterfaceOnAbstractClassFixer/wrong/wrong.php.inc
@@ -1,9 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-
-abstract class SomeClass implements SomeSpecialInterface
-{
-
-}
