+++ packages/EasyCodingStandard/config/symplify.neon
@@ -31,7 +31,6 @@ checkers:
     - Symplify\CodingStandard\Fixer\Naming\PropertyNameMatchingTypeFixer
 
     # Architecture
-    - Symplify\CodingStandard\Fixer\Solid\NoInterfaceOnAbstractClassFixer
     - Symplify\CodingStandard\Fixer\Solid\FinalInterfaceFixer
     - Symplify\CodingStandard\Sniffs\DependencyInjection\NoClassInstantiationSniff
